requirements.txt
pandas==2.2.0
python-dotenv==1.0.1
jupyterlab==4.1.0
matplotlib==3.8.0
